refactor(tests): integrate with centralized kiss_repository tests
